+++ hooks/getContactDetails.js
@@ -6,6 +6,24 @@ import {
 } from "@/assets/Icons8";
 import { Fragment } from "react";
 
+export function getFullName(contact) {
+  const { firstName, middleName, lastName } = contact;
+
+  let displayName = "";
+
+  if (firstName) {
+    displayName += firstName;
+  }
+  if (middleName) {
+    displayName += " " + middleName;
+  }
+  if (lastName) {
+    displayName += " " + lastName;
+  }
+
+  return displayName.trim();
+}
+
 export function getSocialMediaIcon(platform) {
   if (platform === "Instagram") {
     return materialInstagram;
